Reworked logic to lay out things on the figure.
The code that determines where to put things on the figure has now been put in function "lay_out_page", in order to better separate it from the code that actually paints things. The position and extents of each object on the page are now enclosed in a "rect_t" structure. This should allow to improve the lay out of the figure more easily. (Although so far the result is identical to the figures produced before this commit.)