A bug found when reading FITS files has been fixed.
Every time a map containing unseen pixels was read by "hpix_load_fits_component_from_fitsptr", there was a SEGFAULT. This was caused by passing NULL to the "anynull" parameter of fits_read_col_dbl: the CFITSIO manual does not say a word about it, but it seems that the function does not check if the pointer is NULL or not.

The body of the function has been simplified.