New API for hpix_resolution_t
The old way of dealing with hpix_resolution_t was difficult to use with FFIs, as it required to access every member of the structure. The API has changed so that now hpix_resolution_t is dynamically allocated on the heap, and helper functions provide access to its most important members. Beware that this commit breaks API compatibility!